Migrate outRec to Data Locality pattern
